WIP Adding static analyzers
cppcheck is working, but clang-tidy is... special.

Can't really use the cmake functions for integrating those tools,
because I want to store the output of the tools in files. The problem is
that cmake invokes the tools on an file per file basis, which means that
the output files would get overridden by the next compiled file.

cppcheck is intelligent enough to get the needed input files out of the
compile_commands.json file, but clang-tidy is not.