New uC, ditched FTDI chip, and more!
Need to improve the schematic a little bit more.
For example, I want to be able to turn the LED Drivers off using
a MOSFET.

For the PCB Layout: I want to move as many components as possible
to the top layer, so that they are visible when in the case.
Also, I have to consider solderability of the components, the more
complex should all be on one layer for easier soldering.

Next up: Finishing the schematic, improving component placement,
and routing!